vfat first step

+++ user/include/SyscallLib.h
@@ -79,10 +79,10 @@ static inline int mkdir(const char* path, int mode) {
 }
 
 static inline int mount(const char *special, const char *dir, const char *fstype, unsigned long flags, const void *data) {
-    return msyscall(SYSCALL_MOUNT, special, dir, fstype, flags, data, 0);
+    return msyscall(SYSCALL_MOUNT, (u64)special, (u64)dir, (u64)fstype, flags, (u64)data, 0);
 }
 
 static inline int read(int fd, const char *buf, int n) {
-    return msyscall(SYSCALL_READ, fd, buf, n, 0, 0, 0);
+    return msyscall(SYSCALL_READ, fd, (u64)buf, n, 0, 0, 0);
 }
 #endif
